Crate hello_rustfest

Thank you for going to a website you found on a coffee cup, we appreciate it a lot! We hope you are having a great time in Barcelona!

We are The Techno Creatives

The Techno Creatives is a consultancy from Gothenburg, Sweden. We work on a lot of different projects focussed on delivering new and exciting digital experiences. Often, that means we are free to pick whichever technology works best for us -- and we've been able to pick Rust a lot of times recently.

Enums

Role
The dev jobs we are currently hiring for.

Functions

apply_for_a_job
Join us!

hire_us
Does your company want to get started with Rust and/or needs some additional people to get going quickly?