Function hello_rustfest::apply_for_a_job

pub async fn apply_for_a_job<'a>(
    you: &'a Person, 
    job: Role
) -> Result<Interview, Error>

Join us!

At Techno Creatives you will not be joining a boring company where every day is the same. We are a group of people all working with the same goal to help our partners build the best products out there. Our team is always on the lookout for new designers, business developers and business strategist. Don’t be shy and drop a line!

See our jobs page (or browse through these docs and their code) to learn more!